v0.7.0: 機能追加・改善・コード整理
## 新機能
- AutoSongメニュー新設・オートリピート機能
- シーケンス演奏のパート操作モード（Auto/Manual）
- AutoSongモード演奏中のユーザー操作無効化（開始トリガーのみ許可）
- isSequenceActiveMode()によるシーケンス有効モード判定

## 改善
- OTA HTTPクライアントのバッファ溢れ対策（リダイレクト手動解決）
- OTA HTTPS対応（証明書検証スキップ）
- PSRAM確保失敗時のOTAエラー通知追加
- snprintfフォーマット警告修正

## コード整理
- task_wifi: 未使用AsyncWebServerコード削除（-435行）
- task_http_client: ログをM5_LOG*に統一
- 未使用の証明書・HTML定数・includeを削除

+++ main/menu_data/menu_data_system.inl
@@ -309,7 +309,7 @@ struct mi_system_info_t : public mi_normal_t {
 
   const char* getValueText(void) const override {
     static char buf[16];
-    snprintf(buf, sizeof(buf), "v%d.%d.%d", def::app::app_version_major, def::app::app_version_minor, def::app::app_version_patch);
+    snprintf(buf, sizeof(buf), "v%d.%d.%d", (int)def::app::app_version_major, (int)def::app::app_version_minor, (int)def::app::app_version_patch);
     return buf;
   }
   static constexpr const localize_text_t _selector_text = { "Show QR code", "QRコードを表示" };
